Change patients to new audit format.
