+++ .rubocop.yml
@@ -1,0 +1,6 @@
+"Metrics/LineLength":
+  Max: 140
+"Metrics/MethodLength":
+  Max: 20
+"Style/IfUnlessModifier":
+  MaxLineLength: 140
